fixed GCC compile warnings

+++ tutorials/common/image/png.cpp
@@ -130,7 +130,7 @@ namespace embree
     }
 
     // set the individual row_pointers to point at the correct offsets of image_data
-    for (int i = 0; i < height; ++i)
+    for (size_t i = 0; i < height; ++i)
       row_pointers[height - 1 - i] = (unsigned char*)data + i * rowbytes;
  
     // read the png into image_data through row_pointers
